\name{cs}
\alias{cs}
\title{Compactness / Separation Ratio}
\usage{
cs(x, u, v, m, t=NULL, eta, tidx="f")
}
\arguments{
  \item{x}{an object of class \sQuote{ppclust} containing the clustering results from a fuzzy clustering algorithm in the package \pkg{ppclust}. Alternatively, a numeric data frame or matrix containing the data set.}
  \item{u}{a numeric data frame or matrix containing the fuzzy membership values. It should be specified if \code{x} is not an object of \sQuote{ppclust}.}
  \item{v}{a numeric data frame or matrix containing the cluster prototypes. It should be specified if \code{x} is not an object of \sQuote{ppclust}.}
  \item{t}{a numeric data frame or matrix containing the cluster prototypes. It should be specified if \code{x} is not an object of \sQuote{ppclust} and the option \option{e} or \option{g} is assigned to \code{tidx}.}
  \item{m}{a number specifying the fuzzy exponent. It should be specified if \code{x} is not an object of \sQuote{ppclust}.}
  \item{eta}{a number specifying the typicality exponent. It should be specified if \code{x} is not an object of \sQuote{ppclust} and \code{tidx} is either \option{e} or \option{g}.}
  \item{tidx}{a character specifying the type of index. The default is \sQuote{f} for fuzzy index. The other options are \sQuote{e} for extended and \sQuote{g} for generalized index.}
}
\description{
Computes the Separation/Compactness (Bensaid et al, 1996) ratio index in order to validate the result of a fuzzy and/or possibilistic clustering analysis.
}
\details{
The validity-guided (re)clustering (VGC) algorithm uses cluster-validity information to guide a fuzzy (re)clustering process toward better solutions. It starts with a partition generated by a soft or fuzzy clustering algorithm. Then it iteratively alters the partition by applying split-and-merge operations to the clusters (Bensaid, 1996). The authors proposed the Compactness / Separation (CS) ratio index in order to validate the results of VGC. The formula of CS index is:

\deqn{I_{CS}=\sum\limits_{j=1}^k \frac{\sum\limits_{i=1}^n u_{ij}^m \; d_{ij}^2}{\sum\limits_{i=1}^n u_{ij} \; \sum\limits_{l=1}^k||\vec{v_j} - \vec{v_l}||^2}}{I_{CS}=\sum\limits_{j=1}^k \frac{\sum\limits_{i=1}^n u_{ij}^m \; d_{ij}^2}{\sum\limits_{i=1}^n u_{ij} \; \sum\limits_{l=1}^k||\vec{v_j} - \vec{v_l}||^2}}

A lower value of \eqn{I_{CS}}{I_{CS}} indicates a better clustering.
}

\value{
 \item{cs}{CS index value if \code{tidx} is \sQuote{f}}
 \item{cs.e}{extended CS index value if \code{tidx} is \sQuote{e}}
 \item{cs.g}{generalized CS index value if \code{tidx} is \sQuote{g}}
}
\references{
Bensaid, A.M., Hall, L.O., Bezdek, J.C., Clarke, L.P., Silbiger, M.L., Arrington,  J.A. & Murtagh, R.F. (1996). Validity-guided (re)clustering with applications to image segmentation, \emph{IEEE Transactions on Fuzzy Systems}, 4(2):112-123. <10.1109/91.493905>
}

\author{Zeynel Cebeci}

\seealso{
 \code{\link{allindexes}},
 \code{\link{apd}},
 \code{\link{cl}},
 \code{\link{cwb}},
 \code{\link{fhv}},
 \code{\link{fs}},
 \code{\link{kpbm}},
 \code{\link{kwon}},
 \code{\link{mcd}},
 \code{\link{mpc}},
 \code{\link{pbm}},
 \code{\link{pc}},
 \code{\link{pe}},
 \code{\link{si}},
 \code{\link{tss}},
 \code{\link{ws}},
 \code{\link{xb}}
}

\examples{
# Load the dataset iris and use the first four feature columns 
data(iris)
x <- iris[,1:4]

# Run FCM algorithm in the package ppclust 
res.fcm <- ppclust::fcm(x, centers=3)

# Compute the CS index using res.fcm, which is a ppclust object
idx <- cs(res.fcm)
print(idx)
 
# Compute the CS index using X, U and V matrices
idx <- cs(res.fcm$x, res.fcm$u, res.fcm$v)
print(idx)

# Run UPFCM algorithm in the package ppclust 
res.upfc <- ppclust::upfc(x, centers=3)
# Compute the generalized CS index using res.upfc, which is a ppclust object
idx <- cs(res.upfc, tidx="g")
print(idx)
}

\concept{internal validity indexes}
\concept{fuzzy validity indexes}
\concept{validation of clustering}
\concept{cluster validity}
\concept{cluster analysis}

\keyword{cluster}
\keyword{multivariate}