\name{cwb}
\alias{cwb}
\title{Composed Within and Between Scattering Index}
\usage{
cwb(x, u, v, m, t=NULL, eta, av=1, tidx="f")
}
\arguments{
  \item{x}{an object of class \sQuote{ppclust} containing the clustering results from a fuzzy clustering algorithm in the package \pkg{ppclust}. Alternatively, a numeric data frame or matrix containing the data set.}
  \item{u}{a numeric data frame or matrix containing the fuzzy membership values. It should be specified if \code{x} is not an object of \sQuote{ppclust}.}
  \item{v}{a numeric data frame or matrix containing the cluster prototypes. It should be specified if \code{x} is not an object of \sQuote{ppclust}.}
  \item{t}{a numeric data frame or matrix containing the cluster prototypes. It should be specified if \code{x} is not an object of \sQuote{ppclust} and the option \option{e} or \option{g} is assigned to \code{tidx}.}
  \item{m}{a number specifying the fuzzy exponent. It should be specified if \code{x} is not an object of \sQuote{ppclust}.}
  \item{eta}{a number specifying the typicality exponent. It should be specified if \code{x} is not an object of \sQuote{ppclust} and \code{tidx} is either \option{e} or \option{g}.}
  \item{av}{a number specifying the weighting factor which is needed in order to counterbalance both terms of the index value in a proper way. It equals to \eqn{Dis(k_{max})}{Dis(k_{max})}. The default value is 1.}
  \item{tidx}{a character specifying the type of index. The default is \sQuote{f} for fuzzy index. The other options are \sQuote{e} for extended and \sQuote{g} for generalized index.}
}
\description{
Computes the CWB index (Rezaee et al, 1998) in order to validate the result of a fuzzy and/or possibilistic cluster analysis.
}
\details{
The Composed Within and Between Scattering (CWB) index which has been defined by Rezaee et al (1998) is a combined validation index of the average compactness and separation of fuzzy partitions generated by the fuzzy c-means algorithm. The formula of this index as follows:

\deqn{I_{CWB}=\alpha \; Scat(k) + Dis(k)}{I_{CWB}=\alpha \; Scat(k) + Dis(k)}

The first term of the index \eqn{I_{CWB}}{I_{CWB}} indicates the average of the scattering variation within the clusters for \eqn{k}{k}, number of clusters, and it is defined as follows:

\deqn{Scat(k) = \frac{\frac{1}{k} \sum\limits_{j=1}^k ||\sigma(v_i)||}{||\sigma(X)||}}{Scat(k) = \frac{\frac{1}{k} \sum\limits_{j=1}^k ||\sigma(v_i)||}{||\sigma(X)||}},
where \eqn{||x||=\sqrt{x^T x}}{||x||=\sqrt{x^T x}}.

In the above equation, \eqn{\sigma_{v_j}}{\sigma_{v_j}} is the fuzzy variation of the cluster \eqn{j}{j} for the feature \eqn{p}{p}. 

\deqn{\sigma_{v_j}^p=\frac{1}{n} \sum\limits_{i=1}^n u_{ij}(\vec{x}_{i}^p - \vec{v}_{j}^p)^2}{\sigma_{v_j}^p=\frac{1}{n} \sum\limits_{i=1}^n u_{ij}(\vec{x}_{i}^p - \vec{v}_{j}^p)^2}

The second term of the index \eqn{I_{CWB}}{I_{CWB}} indicates the total scattering separation between the clusters. Generally, this term will increase with the number of clusters and is influenced by the geometry of the cluster centers ((Rezaee et al, 1998). It is defined as follows:

\deqn{Dis(k) = \frac{D_{max}}{D_{min}} \sum\limits_{j=1}^k \Big(\sum\limits_{l=1}^k ||v_j - v_l|| \Big)^{-1}}{Dis(k) = \frac{D_{max}}{D_{min}} \sum\limits_{j=1}^k \Big(\sum\limits_{l=1}^k ||v_j - v_l|| \Big)^{-1}}, 

where \eqn{D_{max}=\max{(||v_j - v_l||)}}{D_{max}=\max{(||v_j - v_l||)}} and \eqn{D_{min}=\min{(||v_j - v_l||)}}{D_{min}=\min{(||v_j - v_l||)}} 

The optimal clustering is obtained at the minimum value of \eqn{I_{CWB}}{I_{CWB}}.
}

\value{
 \item{cwb}{CWB index value, if \code{tidx} is \sQuote{f}}
 \item{cwb.e}{extended CWB index, if \code{tidx} is \sQuote{e}}
 \item{cwb.g}{generalized CWB index value, if \code{tidx} is \sQuote{g}}
}
\references{
Rezaee, M. R., Lelieveldt, B. P. & Reiber, J. H. (1998). A new cluster validity index for the fuzzy c-mean. \emph{Pattern Recognition Letters}, 19(3):237-246.<doi:10.1016/S0167-8655(97)00168-2>
}

\author{Zeynel Cebeci}

\seealso{
 \code{\link{allindexes}},
 \code{\link{apd}},
 \code{\link{cl}},
 \code{\link{cs}},
 \code{\link{fhv}},
 \code{\link{fs}},
 \code{\link{kpbm}},
 \code{\link{kwon}},
 \code{\link{mcd}},
 \code{\link{mpc}},
 \code{\link{pbm}},
 \code{\link{pc}},
 \code{\link{pe}},
 \code{\link{sc}},
 \code{\link{si}},
 \code{\link{tss}},
 \code{\link{ws}},
 \code{\link{xb}}
}

\examples{
# Load the dataset iris
data(iris)
x <- iris[,1:4]

# Run FCM algorithm in the package ppclust 
res.fcm <- ppclust::fcm(x, centers=3)

# Compute the CWB index using res.fcm, which is a ppclust object
idx <- cwb(res.fcm)
print(idx)
 
# Compute the CWB index using X, U and V matrices
idx <- cwb(res.fcm$x, res.fcm$u, res.fcm$v)
print(idx)

# Run UPFC algorithm of the package ppclust 
res.upfc <- ppclust::upfc(x, centers=3)
# Compute the generalized CWB index using res.upfc, which is a ppclust object
idx <- cwb(res.upfc, tidx="g")
print(idx)
}

\concept{internal validity indexes}
\concept{fuzzy validity indexes}
\concept{validation of clustering}
\concept{cluster validity}
\concept{cluster analysis}

\keyword{cluster}
\keyword{multivariate}
